refactor: Remove unused components and update imports
This commit removes the Button, ClothingTag, and WeatherKeyInfo components, as they are no longer used in the application. The imports and references to these components have also been updated accordingly. This cleanup improves code organization and reduces unnecessary dependencies.
