+++ src/components/WthrComponents/WthrStats.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-export default function WeatherKeyInfo({ svg, text }) {
+export default function WthrStats({ svg, text }) {
   return (
     <div className="flex justify-center w-full gap-2 px-1 py-2 rounded-lg xl:px-2 xl:py-3 2xl:px-4 2xl:py-4 bg-background dark:bg-dark-background text-primary-tw dark:text-dark-primary-tw">
       <img
@@ -14,6 +14,6 @@ export default function WeatherKeyInfo({ svg, text }) {
   );
 }
 
-WeatherKeyInfo.propTypes = {
+WthrStats.propTypes = {
   text: PropTypes.node.isRequired,
 };
